requirements.txt
# File Search Engine - Python Dependencies
# All versions pinned for stability

# Web Framework
fastapi==0.111.0
uvicorn==0.30.6
python-multipart==0.0.18

# Vector Search
faiss-cpu==1.12.0
sentence-transformers==3.0.1
langchain-huggingface==1.2.0

# LLM Integration
langchain==1.2.7
langchain-community==0.4.1
langchain-text-splitters==1.1.0
langchain-google-genai==4.2.0
langchain-anthropic==1.3.1
langchain-openai==1.1.7
llama-cpp-python==0.2.90
langchain-core==1.2.7

# Document Processing
pypdf==4.3.1
python-docx==1.2.0
python-pptx==1.0.0
openpyxl==3.1.5

# Utilities
psutil==6.1.1
watchdog==6.0.0

scikit-learn==1.6.0
rank-bm25==0.2.2
